Deleting old files
Now we are on the amazon store, I am removing some old files that we don't need anymore like older python repositiories that are not required anymore. I am still looking into the possibility of adding a database for leaderboard functions but it will be alot more complicated to sort through.